% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers.R
\name{augment.casecross}
\alias{augment.casecross}
\title{Augment}
\usage{
\method{augment}{casecross}(x, ...)
}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics]{augment}}, \code{\link[generics]{glance}}, \code{\link[generics]{tidy}}}
}
}
