% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.nsCosinor.R
\name{autoplot.nsCosinor}
\alias{autoplot.nsCosinor}
\title{Plot the trend and seasonal estimates from \code{\link[=nscosinor]{nscosinor()}}}
\usage{
\method{autoplot}{nsCosinor}(object, ...)
}
\arguments{
\item{object}{an \code{nsCosinor} object produced by \code{\link[=nscosinor]{nscosinor()}}.}

\item{...}{unused, for S3 generic compatibility.}
}
\value{
a ggplot object faceted by trend / season(s).
}
\description{
Returns a ggplot of the trend and seasonal components estimated by
\code{\link[=nscosinor]{nscosinor()}}, faceted by component and with a ribbon for the
confidence interval.
}
\examples{
\donttest{
\dontrun{
res <- nscosinor(
  data = CVD, response = "adj", cycles = 12,
  niters = 200, burnin = 100, tau = c(10, 50)
)
autoplot(res)
autoplot(res) + ggplot2::theme_minimal()
}
}
}
\seealso{
\code{\link[=nscosinor]{nscosinor()}}
}
\author{
Nicholas Tierney
}
