% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/third.R
\name{autoplot.third}
\alias{autoplot.third}
\title{Plot the third-order moment from \code{\link[=third]{third()}}}
\usage{
\method{autoplot}{third}(object, ...)
}
\arguments{
\item{object}{a \code{"third"} object produced by \code{\link[=third]{third()}}.}

\item{...}{unused, for S3 generic compatibility.}
}
\value{
a ggplot contour plot.
}
\description{
Produce a ggplot contour of the third-order moment over its
non-redundant region.
}
\examples{
\donttest{
t <- third(CVD$cvd, n.lag = 4)
autoplot(t)
}
}
\seealso{
\code{\link[=third]{third()}}
}
\author{
Nicholas Tierney
}
