\name{cipolygon-internal}
\alias{cipolygon}
\title{Function to draw CI polygon}
\description{Internal function to draw a confidence interval for multiple times as a grey area. For internal use only.}
\usage{cipolygon(time, lower, upper)}
\arguments{
  \item{time}{x-axis.}
  \item{lower}{lower limit of the confidence level.}
  \item{upper}{upper limit of the confidence level.}
}
\author{Adrian Barnett \email{a.barnett@qut.edu.au}}

