% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cipolygon.R
\name{cipolygon}
\alias{cipolygon}
\title{Function to Draw CI Polygon}
\usage{
cipolygon(time, lower, upper)
}
\arguments{
\item{time}{x-axis.}

\item{lower}{lower limit of the confidence level.}

\item{upper}{upper limit of the confidence level.}
}
\description{
Internal function to draw a confidence interval for multiple times as a grey
area. For internal use only.
}
\author{
Adrian Barnett \email{a.barnett@qut.edu.au}
}
