% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosinor.R
\name{cosinor}
\alias{cosinor}
\title{Cosinor Regression Model for Detecting Seasonality in Yearly Data or
Circadian Patterns in Hourly Data}
\usage{
cosinor(
  formula,
  date,
  data,
  family = gaussian(),
  alpha = 0.05,
  cycles = 1,
  rescheck = FALSE,
  type = "daily",
  offsetmonth = FALSE,
  offsetpop = NULL,
  text = TRUE
)
}
\arguments{
\item{formula}{regression formula.}

\item{date}{a date variable if type="daily", or an integer between 1
and 53 if type="weekly", or an integer between 1 and 12 if
type="monthly", or a \link{POSIXct} date if type="hourly".}

\item{data}{data set as a data frame.}

\item{family}{a description of the error distribution and link function to be
used in the model. Available link functions: \code{\link[=gaussian]{gaussian()}} (default),
\code{\link[=identity]{identity()}}, \code{\link[=log]{log()}}, \code{logit()}, \code{cloglog()}. Note, it must have the
parentheses.}

\item{alpha}{Significance level. Default is 0.05.}

\item{cycles}{number of seasonal cycles per year if type="daily",
"weekly" or "monthly"; number of cycles per 24 hours if
type="hourly"}

\item{rescheck}{plot the residual checks (TRUE/FALSE), see
\code{\link[=seasrescheck]{seasrescheck()}}.}

\item{type}{"daily" for daily data (default), or "weekly" for
weekly data, or "monthly" for monthly data, or "hourly" for
hourly data.}

\item{offsetmonth}{include an offset to account for the uneven number of
days in the month (TRUE/FALSE). Should be used for monthly counts (with
\code{family=poisson()}).}

\item{offsetpop}{include an offset for the population (optional), this
should be a variable in the data frame. Do not log-transform the offset as
the log-transform is applied by the function.}

\item{text}{add explanatory text to the returned phase value (TRUE) or
return a number (FALSE). Passed to the \code{invyrfraction} function.}
}
\value{
Returns an object of class "Cosinor" with the following
parts:
\itemize{
\item call: the original call to the cosinor function.
\item glm: an object of class \code{glm} (see \code{\link[=glm]{glm()}}).
\item fitted: fitted values for intercept and cosinor only (ignoring other
independent variables).
\item fitted.plus: standard fitted values, including all other independent
variables.
\item residuals: residuals.
\item date: name of the date variable (in Date format when type = \code{daily}).
}
}
\description{
Fits a cosinor model as part of a generalized linear model.
}
\details{
The cosinor model captures a seasonal pattern using a sinusoid. It is
therefore suitable for relatively simple seasonal patterns that are
symmetric and stationary. The default is to fit an annual seasonal pattern
(\code{cycle}=1), but other higher frequencies are possible (e.g., twice per
year: \code{cycle}=2). The model is fitted using a sine and cosine term that
together describe the sinusoid. These parameters are added to a generalized
linear model, so the model can be fitted to a range of dependent data (e.g.,
Normal, Poisson, Binomial). Unlike the \code{nscosinor} model, the cosinor
model can be applied to unequally spaced data.
}
\examples{

## cardiovascular disease data (offset based on number of days in...
## ...the month scaled to an average month length)
data(CVD)
res = cosinor(cvd~1, date='month', data=CVD, type='monthly',
              family=poisson(), offsetmonth=TRUE)
summary(res)
seasrescheck(res$residuals) # check the residuals
## stillbirth data
data(stillbirth)
res = cosinor(stillborn~1, date='dob', data=stillbirth,
              family=binomial(link='cloglog'))
summary(res)
plot(res)
## hourly indoor temperature data
res = cosinor(bedroom~1, date='datetime', type='hourly', data=indoor)
summary(res)
# to get the p-values for the sine and cosine estimates
summary(res$glm)

}
\references{
Barnett, A.G., Dobson, A.J. (2010) \emph{Analysing Seasonal
Health Data}. Springer.
}
\seealso{
\code{summary.Cosinor}, \code{plot.Cosinor}
}
\author{
Adrian Barnett \email{a.barnett@qut.edu.au}
}
