% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/season-package.R
\docType{data}
\name{CVDdaily}
\alias{CVDdaily}
\title{Daily Cardiovascular Deaths in Los Angeles, 1987--2000}
\format{
A data frame with 5114 observations on the following 16 variables.
\describe{ \item{list("date")}{date of death in date format
(year-month-day)} \item{list("cvd")}{daily number of CVD deaths}
\item{list("dow")}{day of the week (character)} \item{list("tmpd")}{daily
mean temperature (degrees Fahrenheit)} \item{list("o3mean")}{daily mean
ozone (parts per billion)} \item{list("o3tmean")}{daily trimmed mean ozone
(parts per billion)} \item{list("Mon")}{indicator variable for Monday}
\item{list("Tue")}{indicator variable for Tuesday}
\item{list("Wed")}{indicator variable for Wednesday}
\item{list("Thu")}{indicator variable for Thursday}
\item{list("Fri")}{indicator variable for Friday}
\item{list("Sat")}{indicator variable for Saturday}
\item{list("month")}{month (integer from 1 to 12)}
\item{list("winter")}{indicator variable for winter}
\item{list("spring")}{indicator variable for spring}
\item{list("summer")}{indicator variable for summer}
\item{list("autumn")}{indicator variable for autumn} }
}
\source{
From the NMMAPS study.
}
\description{
Daily number of deaths from cardiovascular disease in people aged 75 and
over in Los Angeles for the years 1987 to 2000.
}
\examples{
\donttest{
data(CVDdaily)
plot(CVDdaily$date, CVDdaily$cvd, type='p', xlab='Date',
     ylab='Number of CVD deaths')
}

}
\references{
Samet JM, Dominici F, Zeger SL, Schwartz J, Dockery DW (2000).
\emph{The National Morbidity, Mortality, and Air Pollution Study, Part I:
Methods and Methodologic Issues}. Research Report 94, Health Effects
Institute, Cambridge MA.
}
\keyword{datasets}
