% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_season.R
\name{date_to_season}
\alias{date_to_season}
\title{Add together two numbers.}
\usage{
date_to_season(date)
}
\arguments{
\item{date}{an object lubridate recongnizes as a date}
}
\value{
The season the date is in.
}
\description{
Add together two numbers.
}
\examples{
date_to_season("2018-05-11")
}
