% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers.R
\name{glance.casecross}
\alias{glance.casecross}
\title{Construct a single row summary "glance" of a model, fit, or other object}
\usage{
\method{glance}{casecross}(x, ...)
}
\arguments{
\item{x}{model or other R object to convert to single-row data frame.}

\item{...}{other arguments passed to methods.}
}
\description{
Construct a single row summary "glance" of a model, fit, or other object
}
\examples{
# subset for example
CVDdaily <- subset(CVDdaily, date <= as.Date('1987-12-31'))
# Effect of ozone on CVD death
model1 <- casecross(
  cvd ~ o3mean + tmpd + Mon + Tue + Wed + Thu + Fri + Sat,
  data = CVDdaily
)
summary(model1)
# one-line summary for each model
glance(model1)
}
