% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/season-package.R
\docType{data}
\name{indoor}
\alias{indoor}
\title{Indoor Temperature Data}
\format{
A \code{data.frame} with the following 3 variables.  \describe{
\item{datetime}{date and time in \code{POSIXlt} format}
\item{living}{the living room temperature}
\item{bedroom}{the bedroom temperature} }
}
\source{
Adrian G Barnett.
}
\description{
The data are indoor temperatures (in degrees C) for a bedroom and living
room in a house in Brisbane, Australia for the dates 10 July 2013 to 3
October 2013.  Temperatures were recorded using data loggers which recorded
every hour to the nearest 0.5 degrees.
}
\examples{

data(indoor)
res = cosinor(bedroom~1, date='datetime', type='hourly', data=indoor)
summary(res)

}
\keyword{datasets}
