% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthmean.R
\name{monthmean}
\alias{monthmean}
\title{Calculate monthly mean or adjusted monthly mean for count data.}
\usage{
monthmean(
  data,
  resp,
  offsetpop = NULL,
  adjmonth = c("none", "thirty", "average")
)
}
\arguments{
\item{data}{Data frame with variables "month" and "year"}

\item{resp}{Response variable in data for which the means will be calculated.}

\item{offsetpop}{optional population, used as an offset (default=NULL).}

\item{adjmonth}{adjust monthly counts and scale to a 30 day month ("thirty")
or the average month length ("average") (default="none").}
}
\value{
Returns an object of class "Monthmean" with the following
parts:
\itemize{
\item mean: a vector of length 12 with the monthly means.
}
}
\description{
For time series recorded at monthly intervals it is often useful to examine
(and plot) the average in each month. When using count data we should adjust
the mean to account for the unequal number of days in the month (e.g., 31 in
January and 28 or 29 in February).
}
\details{
This function assumes that the data set (\code{data}) contains variables for
the year and month called year and month, respectively.
}
\examples{
\donttest{
# cardiovascular disease data
mmean <- monthmean(
  data=CVD,
  resp='cvd',
  offsetpop = expression(pop/100000),
  adjmonth = 'average'
  )
mmean
plot(mmean)

mmean <- monthmean(
  data=CVD,
  resp='cvd',
  offsetpop = expression(pop/100000),
  adjmonth = 'thirty'
  )
}

}
\references{
Barnett, A.G., Dobson, A.J. (2010) \emph{Analysing Seasonal
Health Data}. Springer. \doi{doi:10.1007/978-3-642-10748-1}
}
\seealso{
\code{\link[=plot.Monthmean]{plot.Monthmean()}}
}
\author{
Adrian Barnett \email{a.barnett@qut.edu.au}
}
