\name{nochars-internal}
\alias{nochars}
\title{Remove Letters and Characters from a String}
\description{Remove letters and characters from a string to leave only numbers. Removes all letters (upper and lower case) and the characters \dQuote{.}, \dQuote{(} and \dQuote{)}. For internal use only.}
\usage{nochars(text)}
\arguments{
  \item{text}{text string.}
}
\author{Adrian Barnett \email{a.barnett@qut.edu.au}}
