% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/season-package.R
\name{nochars}
\alias{nochars}
\title{Remove Letters and Characters from a String}
\arguments{
\item{text}{text string.}
}
\description{
Remove letters and characters from a string to leave only numbers. Removes
all letters (upper and lower case) and the characters \dQuote{.}, \dQuote{(}
and \dQuote{)}. For internal use only.
}
\author{
Adrian Barnett \email{a.barnett@qut.edu.au}
}
