% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peri.R
\name{peri}
\alias{peri}
\title{Periodogram}
\usage{
peri(data, adjmean = TRUE, plot = lifecycle::deprecated())
}
\arguments{
\item{data}{a data frame.}

\item{adjmean}{subtract the mean from the series before calculating the
periodogram (default=TRUE).}

\item{plot}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Use \code{\link[=autoplot.peri]{autoplot.peri()}} on the
returned object instead.}
}
\value{
an object of class \code{"peri"} (a list) with the following elements:
\itemize{
\item peri: periodogram, I(\eqn{\omega}).
\item f: frequencies in radians, \eqn{\omega}.
\item c: frequencies in cycles of time, \eqn{2\pi/\omega}.
\item amp: amplitude periodogram.
\item phase: phase periodogram.
}

Pass the result to \link[=autoplot.peri]{autoplot()} to draw the plot.
}
\description{
Estimated periodogram using the fast Fourier transform (\code{fft}).
}
\examples{
\donttest{
p <- peri(CVD$cvd)
autoplot(p)
}

}
\seealso{
\code{\link[=autoplot.peri]{autoplot.peri()}}
}
\author{
Adrian Barnett \email{a.barnett@qut.edu.au}
}
