% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.Cosinor.R
\name{plot.Cosinor}
\alias{plot.Cosinor}
\title{Plot the Results of a Cosinor}
\usage{
\method{plot}{Cosinor}(x, ...)
}
\arguments{
\item{x}{a \code{Cosinor} object produced by \code{\link[=cosinor]{cosinor()}}}

\item{\dots}{additional arguments passed to the sinusoid plot.}
}
\value{
connected line plot of fitted sinusoid object produced by \link{cosinor}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Soft-deprecated in favour of
\code{\link[=autoplot.Cosinor]{autoplot.Cosinor()}}, which returns a ggplot object you can extend with
\code{+}. The base R plot below still works.
}
\details{
The code produces the fitted sinusoid based on the intercept and sinusoid.
The y-axis is on the scale of probability if the link function is
"logit" or "cloglog". If the analysis was based on monthly
data then month is shown on the x-axis. If the analysis was based on daily
data then time is shown on the x-axis.
}
\examples{
## cardiovascular disease data (offset based on number of days in...
## ...the month scaled to an average month length)
res <- cosinor(
  cvd ~ 1,
  date = 'month',
  data = CVD,
  type = 'monthly',
  family = poisson(),
  offsetmonth = TRUE
  )
# Recommended:
autoplot(res)
# Still works, but deprecated:
plot(res)
}
\seealso{
\code{\link[=autoplot.Cosinor]{autoplot.Cosinor()}}, \code{\link[=cosinor]{cosinor()}}, \code{\link[=summary.Cosinor]{summary.Cosinor()}},
\code{\link[=seasrescheck]{seasrescheck()}}
}
\author{
Adrian Barnett \email{a.barnett@qut.edu.au}
}
