% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.monthglm.R
\name{plot.monthglm}
\alias{plot.monthglm}
\title{Plot of Monthly Estimates}
\usage{
\method{plot}{monthglm}(x, alpha = 0.05, ylim = NULL, xlab = "", ylab = "", ...)
}
\arguments{
\item{x}{a \code{monthglm} object produced by \code{\link[=monthglm]{monthglm()}}.}

\item{alpha}{statistical significance level of confidence intervals.}

\item{ylim}{y coordinates ranges (the default is NULL, and the limits are
automatically calculated).}

\item{xlab, ylab}{x and y labels. Defaults is no label: "".}

\item{\dots}{additional arguments passed to \code{\link[=plot]{plot()}}.}
}
\value{
Plot of the estimated from a generalized linear model with a
categorical variable of month.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Soft-deprecated in favour of
\code{\link[=autoplot.monthglm]{autoplot.monthglm()}}, which returns a ggplot object. The base R \code{\link[=plot]{plot()}}
will still work, but we recommend using \code{\link[=autoplot.monthglm]{autoplot.monthglm()}}.
}
\examples{
\donttest{
mmodel <- monthglm(
  formula = cvd ~ 1,
  data = CVD,
  family = poisson(),
  offsetpop = expression(pop / 100000),
  offsetmonth = TRUE,
  refmonth = 6
)
# Recommended:
autoplot(mmodel)
# Still works, but deprecated:
plot(mmodel)
}

}
\seealso{
\code{\link[=autoplot.monthglm]{autoplot.monthglm()}}, \code{\link[=monthglm]{monthglm()}}
}
\author{
Adrian Barnett \email{a.barnett@qut.edu.au}
}
