% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.Monthmean.R
\name{plot.Monthmean}
\alias{plot.Monthmean}
\title{Plot of Monthly Mean Estimates}
\usage{
\method{plot}{Monthmean}(x, ...)
}
\arguments{
\item{x}{a \code{Monthmean} object produced by \code{\link[=monthmean]{monthmean()}}.}

\item{\dots}{additional arguments passed to the plot.}
}
\value{
Connected dot plot of estimated monthly means.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Soft-deprecated in favour of
\code{\link[=autoplot.Monthmean]{autoplot.Monthmean()}}, which returns a ggplot object you can extend
with \code{+}. The base R plot below still works.
}
\examples{
\donttest{
mmean <- monthmean(
  data = CVD,
  resp = 'cvd',
  offsetpop = expression(pop / 100000),
  adjmonth = 'average'
)
mmean
# Recommended:
autoplot(mmean)
# Still works, but deprecated:
plot(mmean)
}
}
\seealso{
\code{\link[=autoplot.Monthmean]{autoplot.Monthmean()}}, \code{\link[=monthmean]{monthmean()}}
}
\author{
Adrian Barnett \email{a.barnett@qut.edu.au}
}
