% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.nonlintest.R
\name{plot.nonlintest}
\alias{plot.nonlintest}
\title{Plot the Results of the Non-linear Test}
\usage{
\method{plot}{nonlintest}(x, plot = lifecycle::deprecated(), ...)
}
\arguments{
\item{x}{a \code{nonlintest} object produced by \code{\link[=nonlintest]{nonlintest()}}.}

\item{plot}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} display plot (TRUE) or
return plot (FALSE).}

\item{\dots}{additional arguments to \code{\link[=plot]{plot()}}}
}
\value{
ggplot Contour plot of the region of the third-order moment outside
the test limits generated by \code{\link[=nonlintest]{nonlintest()}}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Soft-deprecated in favour of
\code{\link[=autoplot.nonlintest]{autoplot.nonlintest()}}, which returns a ggplot object you can extend
with \code{+}. The base R-style plot below still works.
}
\examples{
\donttest{
test_res <- nonlintest(data = CVD$cvd, n.lag = 4, n.boot = 1000)
# Recommended:
autoplot(test_res)
# Still works, but deprecated:
plot(test_res)
}
}
\seealso{
\code{\link[=autoplot.nonlintest]{autoplot.nonlintest()}}, \code{\link[=nonlintest]{nonlintest()}}
}
\author{
Adrian Barnett \email{a.barnett@qut.edu.au}
}
