% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCircle.R
\name{plotCircle}
\alias{plotCircle}
\title{Circular Plot}
\usage{
plotCircle(months, dp = 1, ...)
}
\arguments{
\item{months}{monthly variable to plot, the shades of grey of the 12
segments are proportional to this variable. The first result is assumed to
be January, the second February, and so on.}

\item{dp}{decimal places for statistics, default=1.}

\item{\dots}{additional arguments to \code{\link[=plot]{plot()}}}
}
\value{
a donut-type plot of a monthly variable
}
\description{
Circular plot of a monthly variable. This circular plot can be useful for
estimates of an annual seasonal pattern. Darker shades of grey correspond to
larger numbers.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Soft-deprecated in favour of
\code{\link[=plot_circle]{plot_circle()}}, which returns a ggplot object you can extend with \code{+}.
The base R plot below still works.
}
\examples{
\donttest{
# Recommended:
plot_circle(months = seq(1, 12, 1), dp = 0)
# Still works, but deprecated:
plotCircle(months = seq(1, 12, 1), dp = 0)
}

}
\seealso{
\code{\link[=plot_circle]{plot_circle()}}
}
\author{
Adrian Barnett \email{a.barnett@qut.edu.au}
}
