% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.casecross.R
\name{print.casecross}
\alias{print.casecross}
\title{Print the Results of a Case-Crossover Model}
\usage{
\method{print}{casecross}(x, ...)
}
\arguments{
\item{x}{a \code{casecross} object produced by \code{\link[=casecross]{casecross()}}.}

\item{\dots}{optional arguments to \code{\link[=print]{print()}} or \code{\link[=plot]{plot()}} methods.}
}
\value{
printed output of \code{\link[=casecross]{casecross()}}.
}
\description{
The default print method for a \code{casecross} object produced by
\code{\link[=casecross]{casecross()}}.
}
\details{
Uses \code{\link[survival:print.coxph]{print.coxph()}}.
}
\examples{
\donttest{
  CVDdaily <- subset(CVDdaily, date<=as.Date('1987-12-31')) # subset for example
  # Effect of ozone on CVD death
  model1 <- casecross(cvd ~ o3mean+tmpd+Mon+Tue+Wed+Thu+Fri+Sat, data=CVDdaily)
  model1
}
}
\seealso{
\code{\link[=casecross]{casecross()}}, \code{\link[=summary.casecross]{summary.casecross()}}, \code{\link[survival:coxph]{coxph()}}
}
\author{
Adrian Barnett \email{a.barnett@qut.edu.au}
}
