% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.Cosinor.R
\name{print.Cosinor}
\alias{print.Cosinor}
\title{Print the Results of a Cosinor}
\usage{
\method{print}{Cosinor}(x, ...)
}
\arguments{
\item{x}{a \code{Cosinor} object produced by \code{\link[=cosinor]{cosinor()}}.}

\item{\dots}{optional arguments to \code{\link[=print]{print()}} or \code{\link[=plot]{plot()}} methods.}
}
\value{
Prints basic results from \code{\link[=cosinor]{cosinor()}}, using print method for \code{glm}.
}
\description{
The default print method for a \code{Cosinor} object produced by
\code{\link[=cosinor]{cosinor()}}. Uses the \code{glm} method for \code{\link[=print]{print()}}.
}
\examples{
\donttest{
  res <- cosinor(
    cvd ~ 1,
    date = 'month',
    data = CVD,
    type = 'monthly',
    family = poisson(),
    offsetmonth = TRUE
    )
  res
}
}
\seealso{
\code{\link[=cosinor]{cosinor()}} \code{\link[=summary.Cosinor]{summary.Cosinor()}} \code{\link[=glm]{glm()}}
}
\author{
Adrian Barnett \email{a.barnett@qut.edu.au}
}
