% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.Monthmean.R
\name{print.Monthmean}
\alias{print.Monthmean}
\title{Print the Results from Monthmean}
\usage{
\method{print}{Monthmean}(x, digits = 1, ...)
}
\arguments{
\item{x}{a \code{Monthmean} object produced by \code{\link[=monthmean]{monthmean()}}.}

\item{digits}{minimal number of significant digits, see \code{\link[=print.default]{print.default()}}}

\item{\dots}{additional arguments passed to \code{\link[=print]{print()}}.}
}
\value{
a table of values of Month and means
}
\description{
Print the monthly mean estimates from a \code{Monthmean} object produced by
\code{\link[=monthmean]{monthmean()}}.
}
\examples{
\dontttest{
mmean <- monthmean(
  data=CVD,
  resp='cvd',
  offsetpop = expression(pop/100000),
  adjmonth = 'average'
  )
mmean
}
}
\seealso{
\code{monthmean}
}
\author{
Adrian Barnett \email{a.barnett@qut.edu.au}
}
