% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.nonlintest.R
\name{print.nonlintest}
\alias{print.nonlintest}
\title{Print the Results of the Non-linear Test}
\usage{
\method{print}{nonlintest}(x, ...)
}
\arguments{
\item{x}{a \code{nonlintest} object produced by \code{\link[=nonlintest]{nonlintest()}}.}

\item{\dots}{additional arguments to \code{\link[=plot]{plot()}}}
}
\value{
summary of Results of the Non-linear Test.
}
\description{
The default print method for a \code{nonlintest} object produced by
\code{\link[=nonlintest]{nonlintest()}}.
}
\examples{
\donttest{
\dontrun{
  test_res <- nonlintest(data=CVD$cvd, n.lag=4, n.boot=1000)
  test_res
}
}
}
\seealso{
\code{\link[=nonlintest]{nonlintest()}} \code{\link[=plot.nonlintest]{plot.nonlintest()}}
}
\author{
Adrian Barnett \email{a.barnett@qut.edu.au}
}
