% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.nsCosinor.R
\name{print.nsCosinor}
\alias{print.nsCosinor}
\title{Print the Results of a Non-stationary Cosinor}
\usage{
\method{print}{nsCosinor}(x, ...)
}
\arguments{
\item{x}{a \code{nsCosinor} object produced by [\url{nscosinor()}].}

\item{\dots}{further arguments passed to or from other methods.}
}
\description{
The default print method for a \code{nsCosinor} object produced by
[\url{nscosinor()}].
}
\details{
Prints out the model call, number of MCMC samples, sample size and residual
summary statistics.
}
\examples{
\donttest{
# model to fit an annual pattern to the monthly cardiovascular disease data
f <- 12
tau <- c(10,50)
\dontrun{
  res12 <- nscosinor(
    data = CVD,
    response = 'adj',
    cycles = f,
    niters = 250,
    burnin = 100,
    tau = tau
    )
res12
summary(res12)
}
}
}
\seealso{
[\url{nscosinor()}] \code{\link[=summary.nsCosinor]{summary.nsCosinor()}}
}
\author{
Adrian Barnett \email{a.barnett@qut.edu.au}
}
