% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.monthglm.R
\name{print.summary.monthglm}
\alias{print.summary.monthglm}
\title{Printing a summary of a month.glm}
\usage{
\method{print}{summary.monthglm}(x, ...)
}
\arguments{
\item{x}{a \code{summary.monthglm} object produced by \code{\link[=summary.monthglm]{summary.monthglm()}}.}

\item{\dots}{further arguments passed to or from other methods.}
}
\value{
summary of monthglm results. This includes, number of observations,
and a table of the mean, lower, upper, z value, and p value for each of the
months.
}
\description{
Printing a summary of a month.glm
}
\examples{
mmodel <- monthglm(
  formula = cvd~1,
  data = CVD,
  family = poisson(),
  offsetpop = expression(pop/100000),
  offsetmonth = TRUE
  )
summary(mmodel)
}
