% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.nsCosinor.R
\name{print.summary.nscosinor}
\alias{print.summary.nscosinor}
\title{Print a summary of an \code{\link[=nscosinor]{nscosinor()}} object}
\usage{
\method{print}{summary.nscosinor}(x, ...)
}
\arguments{
\item{x}{a \code{summary.nsCosinor} object produced by \code{\link[=summary.nsCosinor]{summary.nsCosinor()}}.}

\item{\dots}{further arguments passed to or from other methods.}
}
\description{
Print a summary of an \code{\link[=nscosinor]{nscosinor()}} object
}
