\name{rinvgamma-internal}
\alias{rinvgamma}
\title{Random inverse gamma distribution}
\description{Internal function to simulate a value from an inverse Gamma distribution, used by \code{nscosinor}. See the MCMCpack library. For internal use only.}
\usage{rinvgamma(n, shape, scale=1)}
\arguments{
  \item{n}{number of observations.}
  \item{shape}{Gamma shape parameter.}
  \item{scale}{Gamma scale parameter (default=1).}
}

