% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/season-package.R
\name{rinvgamma}
\alias{rinvgamma}
\title{Random Inverse Gamma Distribution}
\arguments{
\item{n}{number of observations.}

\item{shape}{Gamma shape parameter.}

\item{scale}{Gamma scale parameter (default=1).}
}
\description{
Internal function to simulate a value from an inverse Gamma distribution,
used by \code{nscosinor}. See the MCMCpack library. For internal use only.
}
