% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.monthglm.R
\name{summary.monthglm}
\alias{summary.monthglm}
\title{Summary for a Monthglm}
\usage{
\method{summary}{monthglm}(object, ...)
}
\arguments{
\item{object}{a \code{monthglm} object produced by \code{nscosinor}.}

\item{\dots}{further arguments passed to or from other methods.}
}
\value{
a list with the following elements:
\itemize{
\item n: sample size.
\item month.ests: parameter estimates for the intercept and months.
\item month.effect: scale of the monthly effects. \code{RR} for rate ratios,
\code{OR} for odds ratios, or empty otherwise.
}
}
\description{
The default summary method for a \code{monthglm} object produced by
\code{monthglm}.
}
\details{
The estimates are the mean, 95\\% confidence interval, Z-value and associated
p-value (comparing each month to the reference month). If Poisson regression
was used then the estimates are shown as rate ratios. If logistic regression
was used then the estimates are shown as odds ratios.
}
\seealso{
\code{monthglm}, \code{plot.monthglm}
}
\author{
Adrian Barnett \email{a.barnett@qut.edu.au}
}
