% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/third.R
\name{third}
\alias{third}
\title{Third-order Moment}
\usage{
third(
  data,
  n.lag,
  centre = TRUE,
  outmax = TRUE,
  plot = lifecycle::deprecated()
)
}
\arguments{
\item{data}{a vector of equally spaced numeric observations (time series).}

\item{n.lag}{the number of lags, maximum = length of time series.}

\item{centre}{centre series by subtracting mean (default=TRUE).}

\item{outmax}{display the (x,y) lag co-ordinates for the maximum and minimum
values (default=TRUE).}

\item{plot}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Use \code{\link[=autoplot.third]{autoplot.third()}} on the
returned object instead. See examples.}
}
\value{
an object of class \code{"third"} (a list) with the following elements:
\itemize{
\item waxis: the axis from \code{-n.lag} to \code{n.lag}.
\item third: the estimated third order moment in the range -n.lag to n.lag,
including the symmetries.
\item n.lag: the maximum lag.
}

Pass the result to \link[=autoplot.third]{autoplot()} to draw the contour
plot.
}
\description{
Estimated third order moment for a time series.
}
\details{
The third-order moment is the extension of the second-order moment
(essentially the autocovariance). The equation for the third order moment at
lags (j,k) is: \eqn{n^{-1}\sum X_t X_{t+j} X_{t+k}}. The third-order moment
is useful for testing for non-linearity in a time series, and is used by
\code{\link[=nonlintest]{nonlintest()}}.
}
\examples{
\donttest{
t <- third(CVD$cvd, n.lag = 12)
autoplot(t)
}

}
\seealso{
\code{\link[=autoplot.third]{autoplot.third()}}
}
\author{
Adrian Barnett \email{a.barnett@qut.edu.au}
}
