% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers.R
\name{tidy.casecross}
\alias{tidy.casecross}
\title{Tidy a casecross object}
\usage{
\method{tidy}{casecross}(x, ...)
}
\arguments{
\item{x}{A \code{casecross} object returned from \code{\link[survival:coxph]{survival::coxph()}}.}

\item{...}{For \code{tidy()}, additional arguments passed to \code{summary(x, ...)}.
Otherwise ignored.}

\item{exponentiate}{Logical indicating whether or not to exponentiate the
coefficient estimates. This is typical for logistic and multinomial
regressions, but a bad idea if there is no log or logit link. Defaults to
FALSE.}

\item{conf.int}{Logical indicating whether or not to include a confidence
interval in the tidied output. Defaults to FALSE.}

\item{conf.level}{The confidence level to use for the confidence interval
if conf.int = TRUE. Must be strictly greater than 0 and less than 1.
Defaults to 0.95, which corresponds to a 95 percent confidence interval.}
}
\description{
This implements the tidy method for \code{\link[broom:tidy.coxph]{broom::tidy.coxph()}}.
}
\examples{
# subset for example
CVDdaily <- subset(CVDdaily, date <= as.Date('1987-12-31'))
# Effect of ozone on CVD death
model1 <- casecross(
  cvd ~ o3mean + tmpd + Mon + Tue + Wed + Thu + Fri + Sat,
  data = CVDdaily
)
summary(model1)

# works with "broom" style tidiers:
# data frame of estimate, std. error, p-value for each model term
# similar to "summary", but in a dataframe
tidy(model1)
# exponentiate the coefficient estimates
tidy(model1, exponentiate = TRUE)
# include confidence intervals in output
tidy(model1, conf.int = TRUE)
# change confidence interval amount
tidy(model1, conf.int = TRUE)
}
