% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/season-package.R
\docType{data}
\name{AFL}
\alias{AFL}
\title{Australian Football League (AFL) Players' Birthdays for the 2009 Season}
\format{
A list with the following 5 variables. 
 \describe{
\item{month}{integer month (1 to 12)} \item{players}{number
of players born in each month (12 observations)}
\item{expected}{expected number of players born in each month (12
observations)} \item{initials}{player initials (617 observations)}
\item{dob}{date of birth in date format (617 observations;
year-month-day format)} }
}
\source{
Dates of birth from Wikipedia.
}
\usage{
AFL
}
\description{
The data are: a) the monthly frequencies of birthdays and an expected number
based on monthly birth statistics for 1975 to 1991. b) all 617 players'
initials and birthdays (excluding non-Australian born players).
}
\examples{
\donttest{data(AFL)
barplot(AFL$players, names.arg=month.abb)
}

}
\keyword{datasets}
