% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.Monthmean.R
\name{plot.Monthmean}
\alias{plot.Monthmean}
\title{Plot of Monthly Mean Estimates}
\usage{
\method{plot}{Monthmean}(x, ...)
}
\arguments{
\item{x}{a \code{Monthmean} object produced by \code{\link[=monthmean]{monthmean()}}.}

\item{\dots}{additional arguments passed to the plot.}
}
\value{
Connected dot plot of estimated monthly means.
}
\description{
Plots estimated monthly means.
}
\examples{
\donttest{
mmean <- monthmean(
  data=CVD,
  resp='cvd',
  offsetpop = expression(pop/100000),
  adjmonth = 'average'
  )
  mmean
  plot(mmean)
}
}
\seealso{
\code{monthmean}
}
\author{
Adrian Barnett \email{a.barnett@qut.edu.au}
}
