% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.Monthmean.R
\name{plot.Monthmean}
\alias{plot.Monthmean}
\title{Plot of Monthly Mean Estimates}
\usage{
\method{plot}{Monthmean}(x, ...)
}
\arguments{
\item{x}{a \code{Monthmean} object produced by \code{monthmean}.}

\item{\dots}{additional arguments passed to the plot.}
}
\description{
Plots estimated monthly means.
}
\seealso{
\code{monthmean}
}
\author{
Adrian Barnett \email{a.barnett<at>qut.edu.au}
}
