\name{plot.Monthmean}
%%\alias{Monthmean}
\alias{plot.Monthmean}
\title{Plot of Monthly Mean Estimates}
\description{Plots estimated monthly means.}
\usage{
\method{plot}{Monthmean}(x, ...)
}
\arguments{
  \item{x}{a \code{Monthmean} object produced by \code{monthmean}.}
  \item{\dots}{additional arguments passed to the plot.}
}
\seealso{\code{monthmean}}
\author{Adrian Barnett \email{a.barnett<at>qut.edu.au}}
