% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.nonlintest.R
\name{plot.nonlintest}
\alias{plot.nonlintest}
\title{Plot the Results of the Non-linear Test}
\usage{
\method{plot}{nonlintest}(x, plot = TRUE, ...)
}
\arguments{
\item{x}{a \code{nonlintest} object produced by \code{\link[=nonlintest]{nonlintest()}}.}

\item{plot}{display plot (TRUE) or return plot (FALSE)}

\item{\dots}{additional arguments to \code{\link[=plot]{plot()}}}
}
\value{
ggplot Contour plot of the region of the third-order moment outside the
test limits generated by \code{\link[=nonlintest]{nonlintest()}}.
}
\description{
Creates a contour plot of the region of the third-order moment outside the
test limits generated by \code{\link[=nonlintest]{nonlintest()}}.
}
\examples{
\donttest{
  test.res <- nonlintest(data=CVD$cvd, n.lag=4, n.boot=1000)
  plot(test.res)
}
}
\seealso{
\code{\link[=nonlintest]{nonlintest()}}
}
\author{
Adrian Barnett \email{a.barnett@qut.edu.au}
}
