% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.nonlintest.R
\name{plot.nonlintest}
\alias{plot.nonlintest}
\title{Plot the Results of the Non-linear Test}
\usage{
\method{plot}{nonlintest}(x, plot = TRUE, ...)
}
\arguments{
\item{x}{a \code{nonlintest} object produced by \code{nonlintest}.}

\item{plot}{display plot (TRUE) or return plot (FALSE)}

\item{\dots}{additional arguments to \code{plot}}
}
\description{
Creates a contour plot of the region of the third-order moment outside the
test limits generated by \code{nonlintest}.
}
\seealso{
\code{nonlintest}
}
\author{
Adrian Barnett \email{a.barnett<at>qut.edu.au}
}
