\name{plot.nonlintest}
\alias{plot.nonlintest}
\title{Plot the results of the non-linear test}
\description{Creates a \code{wireframe} plot of the region of the third-order moment outside the test limits generated by \code{nonlintest}.}
\usage{
\method{plot}{nonlintest}(x, \dots)
}
\arguments{
  \item{x}{a \code{nonlintest} object produced by \code{nonlintest}.}
  \item{\dots}{additional arguments to \code{plot}}
}
\seealso{\code{nonlintest}}
\author{Adrian Barnett \email{a.barnett<at>qut.edu.au}}

