% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.nsCosinor.R
\name{plot.nsCosinor}
\alias{plot.nsCosinor}
\title{Plot the Results of a Non-stationary Cosinor}
\usage{
\method{plot}{nsCosinor}(x, ...)
}
\arguments{
\item{x}{a \code{nsCosinor} object produced by [\url{nscosinor()}].}

\item{\dots}{further arguments passed to or from other methods.}
}
\value{
a plot of class \code{ggplot}.
}
\description{
Plots the trend and season(s) from a \code{nsCosinor} object produced by
[\url{nscosinor()}].
}
\details{
The code produces the season(s) and trend estimates.
}
\examples{
\dontttest{
# model to fit an annual pattern to the monthly cardiovascular disease data
f <- 12
tau <- c(10,50)
\dontrun{
  res12 <- nscosinor(
    data = CVD,
    response = 'adj',
    cycles = f,
    niters = 200,
    burnin = 100,
    tau = tau
    )
plot(res12)
}
}
}
\seealso{
[\url{nscosinor()}]
}
\author{
Adrian Barnett \email{a.barnett@qut.edu.au}
}
