% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.nsCosinor.R
\name{plot.nsCosinor}
\alias{plot.nsCosinor}
\title{Plot the Results of a Non-stationary Cosinor}
\usage{
\method{plot}{nsCosinor}(x, ...)
}
\arguments{
\item{x}{a \code{nsCosinor} object produced by \code{nscosinor}.}

\item{\dots}{further arguments passed to or from other methods.}
}
\value{
\item{gplot}{A plot of class \code{ggplot}}
}
\description{
Plots the trend and season(s) from a \code{nsCosinor} object produced by
\code{nscosinor}.
}
\details{
The code produces the season(s) and trend estimates.
}
\seealso{
\code{nscosinor}
}
\author{
Adrian Barnett \email{a.barnett<at>qut.edu.au}
}
