% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.casecross.R
\name{print.casecross}
\alias{print.casecross}
\title{Print the Results of a Case-Crossover Model}
\usage{
\method{print}{casecross}(x, ...)
}
\arguments{
\item{x}{a \code{casecross} object produced by \code{casecross}.}

\item{\dots}{optional arguments to \code{print} or \code{plot} methods.}
}
\description{
The default print method for a \code{casecross} object produced by
\code{casecross}.
}
\details{
Uses \code{print.coxph}.
}
\seealso{
\code{casecross}, \code{summary.casecross}, \code{coxph}
}
\author{
Adrian Barnett \email{a.barnett<at>qut.edu.au}
}
