\name{print.casecross}
%%\alias{casecross}
\alias{print.casecross}
\title{Print the results of a case-crossover model}
\description{The default print method for a \code{casecross} object produced by \code{casecross}.}
\usage{
\method{print}{casecross}(x, ...)
}
\arguments{
  \item{x}{a \code{casecross} object produced by \code{casecross}.}
  \item{\dots}{optional arguments to \code{print} or \code{plot} methods.}
}
\details{Uses \code{print.coxph}.}
\seealso{\code{casecross}, \code{summary.casecross}, \code{coxph}}
\author{Adrian Barnett \email{a.barnett<at>qut.edu.au}}

