% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.Cosinor.R
\name{print.Cosinor}
\alias{print.Cosinor}
\title{Print the Results of a Cosinor}
\usage{
\method{print}{Cosinor}(x, ...)
}
\arguments{
\item{x}{a \code{Cosinor} object produced by \code{cosinor}.}

\item{\dots}{optional arguments to \code{print} or \code{plot} methods.}
}
\description{
The default print method for a \code{Cosinor} object produced by
\code{cosinor}.
}
\details{
Uses \code{print.glm}.
}
\seealso{
\code{cosinor}, \code{summary.Cosinor}, \code{glm}
}
\author{
Adrian Barnett \email{a.barnett<at>qut.edu.au}
}
