\name{print.Cosinor}
%%\alias{cosinor}
\alias{print.Cosinor}
\title{Print the results of a cosinor}
\description{The default print method for a \code{Cosinor} object produced by \code{cosinor}.}
\usage{
\method{print}{Cosinor}(x, ...)
}
\arguments{
  \item{x}{a \code{Cosinor} object produced by \code{cosinor}.}
  \item{\dots}{optional arguments to \code{print} or \code{plot} methods.}
}
\details{Uses \code{print.glm}.}
\seealso{\code{cosinor}, \code{summary.Cosinor}, \code{glm}}
\author{Adrian Barnett \email{a.barnett@qut.edu.au}}

