% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.monthglm.R
\name{print.monthglm}
\alias{print.monthglm}
\title{Print \code{monthglm}}
\usage{
\method{print}{monthglm}(x, ...)
}
\arguments{
\item{x}{Object of class \code{monthglm}}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print \code{monthglm}
}
