\name{print.Monthmean}
\alias{print.Monthmean}
\title{Print the results from monthmean}
\description{Print the monthly means from a \code{Monthmean} object produced by \code{monthmean}.}
\usage{
\method{print}{Monthmean}(x,...)
}
\arguments{
  \item{x}{a \code{Monthmean} object produced by \code{monthmean}.}
  \item{\dots}{additional arguments passed to the print.}
}
\details{The code prints the monthly mean estimates.}
\seealso{\code{monthmean}}
\author{Adrian Barnett \email{a.barnett<at>qut.edu.au}}

