% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.nonlintest.R
\name{print.nonlintest}
\alias{print.nonlintest}
\title{Print the Results of the Non-linear Test}
\usage{
\method{print}{nonlintest}(x, ...)
}
\arguments{
\item{x}{a \code{nonlintest} object produced by \code{nonlintest}.}

\item{\dots}{additional arguments to \code{plot}}
}
\description{
The default print method for a \code{nonlintest} object produced by
\code{nonlintest}.
}
\seealso{
\code{nonlintest}, \code{plot.nonlintest}
}
\author{
Adrian Barnett \email{a.barnett<at>qut.edu.au}
}
