\name{print.nonlintest}
\alias{print.nonlintest}
\title{Print the results of the non-linear test}
\description{The default print method for a \code{nonlintest} object
  produced by \code{nonlintest}.}
\usage{
\method{print}{nonlintest}(x, \dots)
}
\arguments{
  \item{x}{a \code{nonlintest} object produced by \code{nonlintest}.}
    \item{\dots}{additional arguments to \code{plot}}
}
\seealso{\code{nonlintest}, \code{plot.nonlintest}}
\author{Adrian Barnett \email{a.barnett<at>qut.edu.au}}

