% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.nsCosinor.R
\name{print.nsCosinor}
\alias{print.nsCosinor}
\title{Print the Results of a Non-stationary Cosinor}
\usage{
\method{print}{nsCosinor}(x, ...)
}
\arguments{
\item{x}{a \code{nsCosinor} object produced by \code{nscosinor}.}

\item{\dots}{further arguments passed to or from other methods.}
}
\description{
The default print method for a \code{nsCosinor} object produced by
\code{nscosinor}.
}
\details{
Prints out the model call, number of MCMC samples, sample size and residual
summary statistics.
}
\seealso{
\code{nscosinor}, \code{summary.nsCosinor}
}
\author{
Adrian Barnett \email{a.barnett@qut.edu.au}
}
