% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.Cosinor.R
\name{print.summary.Cosinor}
\alias{print.summary.Cosinor}
\title{printing a summary of a Cosinor}
\usage{
\method{print}{summary.Cosinor}(x, ...)
}
\arguments{
\item{x}{a \code{summary.Cosinor} object produced by \code{summary.Cosinor}}

\item{\dots}{optional arguments to \code{print} or \code{plot} methods.}
}
\description{
printing a summary of a Cosinor
}
