% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.monthglm.R
\name{print.summary.monthglm}
\alias{print.summary.monthglm}
\title{Printing a summary of a month.glm}
\usage{
\method{print}{summary.monthglm}(x, ...)
}
\arguments{
\item{x}{a \code{summary.monthglm} object produced by \code{\link[=summary.monthglm]{summary.monthglm()}}.}

\item{\dots}{further arguments passed to or from other methods.}
}
\description{
Printing a summary of a month.glm
}
