% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.nsCosinor.R
\name{print.summary.nsCosinor}
\alias{print.summary.nsCosinor}
\title{Print a summary of an \code{\link[=nscosinor]{nscosinor()}} object}
\usage{
\method{print}{summary.nsCosinor}(x, ...)
}
\arguments{
\item{x}{a \code{summary.nsCosinor} object produced by \code{\link[=summary.nsCosinor]{summary.nsCosinor()}}.}

\item{\dots}{further arguments passed to or from other methods.}
}
\value{
summary - Statistics for non-stationary cosinor based on MCMC
chains.
}
\description{
Print a summary of an \code{\link[=nscosinor]{nscosinor()}} object
}
\examples{
\donttest{
# model to fit an annual pattern to the monthly cardiovascular disease data
f <- c(12)
tau <- c(10,50)
\dontrun{
  res12 <- nscosinor(
    data = CVD,
    response = 'adj',
    cycles = f,
    niters = 200,
    burnin = 100,
    tau = tau
    )
summary(res12)
}
}
}
