% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.nscosinor.R
\name{print.summary.nscosinor}
\alias{print.summary.nscosinor}
\title{printing a summary of an nscosinor}
\usage{
\method{print}{summary.nscosinor}(x, ...)
}
\arguments{
\item{x}{a \code{summary.nsCosinor} object produced by
\code{summary.nsCosinor}}

\item{\dots}{further arguments passed to or from other methods.}
}
\description{
printing a summary of an nscosinor
}
