% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.casecross.R
\name{summary.casecross}
\alias{summary.casecross}
\title{Summary of the Results of a Case-crossover Model}
\usage{
\method{summary}{casecross}(object, ...)
}
\arguments{
\item{object}{a \code{casecross} object produced by \code{casecross}.}

\item{\dots}{further arguments passed to or from other methods.}
}
\description{
The default summary method for a \code{casecross} object produced by
\code{casecross}.
}
\details{
Shows the number of control days, the average number of control days per
case days, and the parameter estimates.
}
\seealso{
\code{casecross}
}
\author{
Adrian Barnett \email{a.barnett<at>qut.edu.au}
}
